+++ packages/vidstack/src/global/layouts/default.ts
@@ -1,0 +1,26 @@
+import type { DefaultLayoutProps } from '../../components';
+import type { VidstackPlayerLayoutLoader } from './loader';
+
+export class VidstackPlayerLayout implements VidstackPlayerLayoutLoader {
+  constructor(readonly props?: Partial<DefaultLayoutProps>) {}
+
+  async load() {
+    await import('../../elements/bundles/player-layouts/default');
+    await import('../../elements/bundles/player-ui');
+  }
+
+  create() {
+    const layouts = [
+      document.createElement('media-audio-layout'),
+      document.createElement('media-video-layout'),
+    ];
+
+    if (this.props) {
+      for (const [prop, value] of Object.entries(this.props)) {
+        for (const el of layouts) el[prop] = value;
+      }
+    }
+
+    return layouts;
+  }
+}