+++ packages/vidstack/src/global/layouts/loader.ts
@@ -1,0 +1,4 @@
+export interface VidstackPlayerLayoutLoader {
+  load(): void | Promise<void>;
+  create(): HTMLElement[] | Promise<HTMLElement[]>;
+}