+++ packages/vidstack/src/globals.d.ts
@@ -5,6 +5,7 @@ declare global {
   const __DEV__: boolean;
   const __SERVER__: boolean;
   const __TEST__: boolean;
+  const __CDN__: boolean;
 }
 
 export {};